first pass of docker build for arch
