+++ arm64.yml
@@ -1,0 +1,8 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       args:
+         IMAGE_ARCH: arm64v8
+         QEMU_FILE: input/qemu-aarch64-static
