--- assets/arm/assets.txt
@@ -1,7 +1,0 @@
-busybox 1549996374 aa6097e24e6e1282b1ba7375df2385d0
-libdisableselinux.so 1549996374 9fa7547d76427e5d825a945d64a75b4c
-rootfs.tar.gz.part00 1549996374 70ca044211bd555d0856afb46b442ca9
-rootfs.tar.gz.part01 1549996375 f46747979634e5d60e6d9c1736aed9a1
-rootfs.tar.gz.part02 1549996375 2974a430c83218f1e4d64a00d76bc893
-rootfs.tar.gz.part03 1549996376 1318e03f2718ab7fc0db2f0cba5bdfc4
-rootfs.tar.gz.part04 1549996376 f23b1b6ffff95e519f6d0dd626be8831
