--- assets/arm64/assets.txt
@@ -1,7 +1,0 @@
-busybox 1549996376 de3e816fb517ba1c9ee8997daf1c6d53
-libdisableselinux.so 1549996376 3c357f2fe252a4ed1e79d85514dac6bb
-rootfs.tar.gz.part00 1549996376 b3fa228c2d0fa0a66dba923fe34adb2b
-rootfs.tar.gz.part01 1549996377 99943b43059f8caabbe82688aea703ee
-rootfs.tar.gz.part02 1549996377 7eca452d57799dd4001515d2e6592ec8
-rootfs.tar.gz.part03 1549996378 7d7f82d7dd9355d3b4d0c9be63d027ad
-rootfs.tar.gz.part04 1549996378 b2eaab3dfb9649e34b31f361fbf54786
