+++ assets/ld.so.preload
@@ -1,0 +1,1 @@
+/support/libdisableselinux.so
