--- assets/x86_64/assets.txt
@@ -1,7 +1,0 @@
-busybox 1549996380 8fe0a49a51e979013850824bbd582df2
-libdisableselinux.so 1549996380 68e16e55db932292a5dff639346fd743
-rootfs.tar.gz.part00 1549996381 130827726bbb2ab27ce81028cbf23651
-rootfs.tar.gz.part01 1549996381 7da507531aade5c91568b26adf4b68d5
-rootfs.tar.gz.part02 1549996381 842942065fd6db2ef5460fa31d713b6c
-rootfs.tar.gz.part03 1549996382 af9c2811607bceb76ca555a48957db7d
-rootfs.tar.gz.part04 1549996382 b3bc6e584064194efcb73ab39d19bdd0
