--- scripts/shrinkRootfs.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-# pacman's equivalent of apt clean
-
-pacman -Scc
